frontend/requirements.txt
streamlit==1.30.0
requests==2.31.0
pillow==10.2.0
plotly==5.18.0
streamlit-lottie==0.0.5
